Logical Changes

+++ MCDStorageChest/Logic/GameCalculator.cs
@@ -39,12 +39,12 @@ namespace MCDStorageChest.Logic
             return 100 * (level - 1) * (3 * level - 1);
         }
 
-        public static double characterPowerFromEquippedItemPowers(double melee, double armor, double ranged, double slot1, double slot2, double slot3)
+        public static double characterPowerFromEquippedItemPowers(decimal melee, decimal armor, decimal ranged, decimal slot1, decimal slot2, decimal slot3)
         {
             var weightedAverage
                 = (melee + armor + ranged) / 4
                 + (slot1 + slot2 + slot3) / 12;
-            return weightedAverage;
+            return (double)weightedAverage;
         }
     }
 }
